# docs/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Credity API
  version: 1.1.0
  description: |
    Credity standards-aligned API surface for issuance, presentation, verification,
    status, and asynchronous anchoring.
  x-contract-freeze-date: 2026-02-12
  x-deprecation-policy:
    notice_period_days: 90
    sunset_header_required: true
    communication_channels:
      - changelog
      - dashboard-banner
      - email
servers:
  - url: https://api.credity.in
  - url: http://localhost:5001
    description: Issuer local service
  - url: http://localhost:5003
    description: Recruiter local service
...
paths:
...
  /v1/compliance/consents:
    get:
      tags: [compliance]
      summary: List consent grants
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        "200":
          description: Consent grants
    post:
      tags: [compliance]
      summary: Create consent grant
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentGrantContract"
      responses:
        "201":
          description: Consent grant created
  /v1/compliance/consents/{consentId}/revoke:
    post:
      tags: [compliance]
      summary: Revoke consent grant
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: consentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Consent revoked
        "404":
          description: Consent not found
  /v1/compliance/data-requests:
    get:
      tags: [compliance]
      summary: List data export/delete requests
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        "200":
          description: Data request list
  /v1/compliance/data-requests/export:
    post:
      tags: [compliance]
      summary: Submit data export request
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        "202":
          description: Data export request accepted
  /v1/compliance/data-requests/delete:
    post:
      tags: [compliance]
      summary: Submit data deletion request
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        "202":
          description: Data delete request accepted
  /v1/compliance/certin/incidents:
    get:
      tags: [compliance]
      summary: List CERT-In incident records
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        "200":
          description: CERT-In incidents
    post:
      tags: [compliance]
      summary: Create CERT-In incident record
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        "201":
          description: Incident created
  /v1/compliance/audit-log/export:
    get:
      tags: [compliance]
      summary: Export immutable audit log
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [json, ndjson]
      responses:
        "200":
          description: Audit log export
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ConsentGrantContract:
      type: object
      required:
        - id
        - subject_id
        - verifier_id
        - purpose
        - data_elements
        - expiry
        - consent_proof
      properties:
        id:
          type: string
        subject_id:
          type: string
        verifier_id:
          type: string
        purpose:
          type: string
        data_elements:
          type: array
          items:
            type: string
        expiry:
          type: string
        revocation_ts:
          type: [string, "null"]
        consent_proof:
          type: object
          additionalProperties: true
...